Simplify unflatten code a bit
Such approach is the rudiment from 'offset as a count' way to handle
offsets when we can don't unflatten some tuples until reach the offset.
Now we use 'offset as a value of primary key type' way, so we can use
indexes in the fast case to handle offset, but need to unflatten all
tuples found by an index.