Use gql_type to create top level collections
The commmit introduces the idea that the `Query` object is just a
collection. If so, it is natural to add this type to GraphQL using
the same procedure as for other types.

This is important change, because it not only deletes current code
duplication, but allows to not repeat yourself during further coding.