Fix GraphiQL warning on the pcre argument
The problem was in incorrect schema: sub-arguments of the pcre argument
for non-null strings was generated as non-null that triggers a warning
in GraphiQL in case when the sub-argument is omitted. Graphql-lua
(graphql/core/validate.lua) does not check for this kind of
inconsistency between a schema and a query, so lack of the sub-argument
works as expected in test/common/pcre.test.lua.

The added test with immediate sub-arguments of the pcre argument does
not catch the bug, but I decided to leave it in the commit.

Fixes #156.