Set default timeout to infinity
It is annoying to debug flaky fails in testing here (eps. when there are
several types of flaky fails) and likely will so in other systems,
which'll use graphql. I think a request that hang is more easy to debug,
because it is not so flaky and it is possible to collect runtime
information what is going on.