Fix name clash b/w 1:1 connections and collections
Before we could not use the same name for collection and
1:1 connection (or collection and 1:1 multi-head connection
variant) because this created two different types with same
names in GraphQL schema: Object type for collection and
InputObject type for connection. Now we can use the same names,
since we add 'result_' to the name of GraphQL collection
Object type.

Close #191, close #104