Fix flaky fails with hang tests
The problem is that frontend wait infinitely for a storage server, which
have not been initialized due to the error like the following.

```
Start failed: builtin/box/console.lua:560: failed to create server
localhost:57988: Address already in use
```

The problem is described in [1] (see comment itself, the issue is about
the another problem).

The gist of the fix is using unix sockets for admin console of
non-default servers (storages), it is why test-run update is needed. The
option 'use_unix_sockets' affects only console (admin) socket and does
nothing with binary (listen) port of a storage. It is critical in the
benchmarking test suite.

[1]: https://github.com/tarantool/test-run/issues/115#issuecomment-413850864