Give an empty multihead connection as null
This commit changes behaviour in the following cases:

1. When a union variant to which a connected object belongs is not appear
   in a query, the library returns box.NULL w/o any wrappers.

2. When no connected object is found, the library returns box.NULL as
   well even if all-nulls check before fails (typically because of
   disable_dangling_check option, but also it is possible because of
   using different source fields across multihead connection variants).

The first case is tested with to_avro_multihead.test.lua. The latter one
does not tested.

In these cases we could receive `{"collection_name": null}` as a result
instead of just `null`. Only `null` is valid result for a avro-schema
union.

Both variant however are valid by our avro-schema generated for a query,
because a multi-head connection represented as an object with nullable
fields. There is a reason for that, see the comment for the
union_to_avro() function in graphql/query_to_avro.lua.

Fixes #253.