avro-schema generation: support record inside map
This is more like the quick fix. I think the right way would be create a
module to convert a GraphQL type (w/o selections) into avro-schema.

Also a union inside a map unlikely will work, but here I have fixed only
the case with a record inside a map. I'll follow it up if there will be
a demand.

Fixes #283.